# docs/openapi.yml
openapi: "3.0.3"
info:
  title: Bienenhaus Propiedades API
  description: |
    API RESTful del panel de administración de Bienenhaus Propiedades.
    Inmobiliaria boutique con CRM, tasaciones ACM, portales y redes sociales.
  version: "1.0.0"
  contact:
    name: Bienenhaus Propiedades
    url: https://bienenhaus.com.ar
servers:
  - url: https://bienenhaus.com.ar/api
    description: Producción
  - url: http://localhost:5000/api
    description: Desarrollo
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Lead:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        status: { type: string, enum: [nuevo, contactado, calificado, visita_agendada, visita_realizada, negociacion, cerrado_ganado, cerrado_perdido] }
        agent_name: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
paths:
...
  /crm/leads:
    get:
      summary: Lista leads del CRM
      security: [{ SessionAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: agent_id
          schema: { type: integer }
      responses:
        "200":
          description: Lista paginada de leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Lead" } }
                      total: { type: integer }
    post:
      summary: Crea un lead
      security: [{ SessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string }
      responses:
        "201":
          description: Lead creado
  "/crm/leads/{id}":
    get:
      summary: Obtiene un lead
      security: [{ SessionAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Lead
    delete:
      summary: Elimina un lead
      security: [{ SessionAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Eliminado
...
  /auth/login:
    post:
      summary: Inicia sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Sesión iniciada
        "401":
          description: Credenciales inválidas
...